♻️ Refactor the participants frontend code.
